requirements.txt
websocket-client==1.6.3
python-dotenv==1.0.0
requests==2.32.4
# pandas>=2.0.0
# numpy>=1.24.0
# binance-connector>=3.10.0